standard deviations are now floating point, not integer

+++ Signal/General/SpectralKurtosisCUDA.cu
@@ -57,7 +57,7 @@ void CUDA::SpectralKurtosisEngine::detect_ft (const dsp::TimeSeries* input,
 void CUDA::SpectralKurtosisEngine::detect_fscr (
   const dsp::TimeSeries* input,
   dsp::BitSeries* output,
-  const float mu2, const unsigned std_devs,
+  const float mu2, const float std_devs,
   unsigned schan, unsigned echan)
 
 {
